Moves rate update form, logic and functions to separate component. Clears HostProfile component even more.
